+++ GHEtool/test/unit-tests/test_main_class.py
@@ -793,6 +793,38 @@ def test_gfunction():
     np.testing.assert_array_almost_equal(borefield.gfunction([6000, 60000, 600000]), np.array([0.63751082, 1.70657847, 2.84227252]))
 
 
+def test_gfunction_with_irregular_depth():
+    borefield = Borefield()
+    borefield.set_ground_parameters(data_ground_flux)
+    temp = [
+        gt.boreholes.Borehole(100, 4, 0.075, 0, 0),
+        gt.boreholes.Borehole(150, 4, 0.075, 10, 0),
+        gt.boreholes.Borehole(50, 4, 0.075, 100, 0)
+    ]
+    borefield.borefield = temp
+    assert borefield.H == 100
+    g_values = borefield.gfunction([6000, 60000, 600000])
+    borefield._update_borefield_depth(100)
+    assert not np.array_equal(borefield.gfunction([6000, 60000, 600000]), g_values)
+
+    borefield = Borefield()
+    borefield.set_ground_parameters(data_ground_flux)
+    temp = [
+        gt.boreholes.Borehole(100, 4, 0.075, 0, 0),
+        gt.boreholes.Borehole(150, 4, 0.075, 10, 0),
+        gt.boreholes.Borehole(50, 4, 0.075, 100, 0)
+    ]
+    borefield.borefield = temp
+    assert borefield.H == 100
+    temp = [
+        gt.boreholes.Borehole(100, 4, 0.075, 0, 0),
+        gt.boreholes.Borehole(100, 4, 0.075, 10, 0),
+        gt.boreholes.Borehole(100, 4, 0.075, 100, 0)
+    ]
+    borefield.borefield = temp
+    assert not np.array_equal(borefield.gfunction([6000, 60000, 600000]), g_values)
+
+
 def test_load_duration(monkeypatch):
     borefield = Borefield()
     monkeypatch.setattr(plt, "show", lambda: None)